move all modals to folder, fix files importing them
